Moved alembic to sql_db dir
