+++ services/assistant/app/infra/database/sql_db/alembic/script.py.mako
