+++ services/assistant/app/infra/database/sql_db/alembic/versions/8f81a875074f_initial_migration.py
