+++ services/assistant/app/infra/database/sql_db/alembic/versions/e0fa9bdc69e1_adding_xray_record_table.py
